Moving to a multi-table arrangement.
Simple URI scheme:
/ Root, where you get info on the site, and can make new tables.
/:table Table, where :table is a UUID4.

At a table, you'll be able to blind, hand, and open cards from the
magic list of cards, the Deck. You'll be able to move cards between
those three piles in front of you (blind -> hand -> open, at least),
and you'll be able to discard to the other magic list, the Discard.

When a table is made, it'll be prepopulated with an empty Discard
and a full Deck. Later, I can think about modifying the type of
Deck, but for now, it'll assume playing cards.